backend/requirements.txt
flask
flask-cors
torch
transformers
shap
scikit-learn
pandas
numpy
datasets
tqdm
python-dotenv
accelerate
